+++ Defs/JobDef.xml
@@ -1,32 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Defs>
-  <JobDef>
-    <defName>PrisonLabor_PrisonerSupervise</defName>
-    <driverClass>PrisonLabor.Core.AI.JobDrivers.JobDriver_Supervise</driverClass>
-    <reportString>watching prisoner TargetA.</reportString>
-    <alwaysShowWeapon>true</alwaysShowWeapon>
-  </JobDef>
-  <JobDef>
-    <defName>PrisonLabor_Arrest</defName>
-    <driverClass>JobDriver_TakeToBed</driverClass>
-    <reportString>arresting TargetA.</reportString>
-    <alwaysShowWeapon>true</alwaysShowWeapon>
-    <makeTargetPrisoner>true</makeTargetPrisoner>
-    <casualInterruptible>false</casualInterruptible>
-  </JobDef>
-  <JobDef>
-    <defName>PrisonLabor_Mine_Tweak</defName>
-    <driverClass>PrisonLabor.Tweaks.JobDriver_Mine_Tweak</driverClass>
-    <reportString>digging at TargetA.</reportString>
-  </JobDef>
-  <JobDef>
-    <defName>PrisonLabor_Harvest_Tweak</defName>
-    <driverClass>PrisonLabor.Tweaks.JobDriver_PlantHarvest_Tweak</driverClass>
-    <reportString>harvesting TargetA.</reportString>
-  </JobDef>
-  <JobDef>
-    <defName>PrisonLabor_CutPlant_Tweak</defName>
-    <driverClass>PrisonLabor.Tweaks.JobDriver_PlantCut_Tweak</driverClass>
-    <reportString>cutting TargetA.</reportString>
-  </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_PrisonerSupervise</defName>
+        <driverClass>PrisonLabor.Core.AI.JobDrivers.JobDriver_Supervise</driverClass>
+        <reportString>watching prisoner TargetA.</reportString>
+        <alwaysShowWeapon>true</alwaysShowWeapon>
+    </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_PrisonerDeliverFoodSupervise</defName>
+        <driverClass>PrisonLabor.Core.AI.JobDrivers.JobDriver_JailorDeliverFood</driverClass>
+        <reportString>Delivering food to TargetA.</reportString>
+        <alwaysShowWeapon>true</alwaysShowWeapon>
+    </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_Arrest</defName>
+        <driverClass>JobDriver_TakeToBed</driverClass>
+        <reportString>arresting TargetA.</reportString>
+        <alwaysShowWeapon>true</alwaysShowWeapon>
+        <makeTargetPrisoner>true</makeTargetPrisoner>
+        <casualInterruptible>false</casualInterruptible>
+    </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_Mine_Tweak</defName>
+        <driverClass>PrisonLabor.Tweaks.JobDriver_Mine_Tweak</driverClass>
+        <reportString>digging at TargetA.</reportString>
+    </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_Harvest_Tweak</defName>
+        <driverClass>PrisonLabor.Tweaks.JobDriver_PlantHarvest_Tweak</driverClass>
+        <reportString>harvesting TargetA.</reportString>
+    </JobDef>
+    <JobDef>
+        <defName>PrisonLabor_CutPlant_Tweak</defName>
+        <driverClass>PrisonLabor.Tweaks.JobDriver_PlantCut_Tweak</driverClass>
+        <reportString>cutting TargetA.</reportString>
+    </JobDef>
 </Defs>
