+++ Defs/ThinkTreeDef.xml
@@ -1,26 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Defs>
-  <ThinkTreeDef>
-      <defName>PrisonLabor_WorkThinkTree</defName>
-      <insertTag>Humanlike_PostDuty</insertTag>
-      <insertPriority>80</insertPriority>
-      <thinkRoot Class="PrisonLabor.Core.AI.ThinkNodes.ThinkNode_Labor">
-      	<invert>false</invert>
-      	<subNodes>
-          <li Class="PrisonLabor.Core.AI.ThinkNodes.ThinkNode_SeekSafeTemperature">
+    <ThinkTreeDef>
+        <defName>PrisonLabor_WorkThinkTree</defName>
+        <insertTag>Humanlike_PostDuty</insertTag>
+        <insertPriority>80</insertPriority>
+        <thinkRoot Class="PrisonLabor.Core.AI.ThinkNodes.ThinkNode_Labor">
+            <invert>false</invert>
             <subNodes>
-              <li Class="JobGiver_SeekSafeTemperature"/>
+                <li Class="PrisonLabor.Core.AI.ThinkNodes.ThinkNode_SeekSafeTemperature">
+                    <subNodes>
+                        <li Class="JobGiver_SeekSafeTemperature" />
+                    </subNodes>
+                </li>
+                <li Class="ThinkNode_Priority">
+                    <subNodes>
+                        <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_BedTime" />
+                        <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_PickupWeapon" />
+                        <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_Diet" />
+                        <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_Labor" />
+                    </subNodes>
+                </li>
             </subNodes>
-          </li>
-      		<li Class="ThinkNode_Priority">
-      			<subNodes>
-              <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_BedTime" />
-              <li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_PickupWeapon" />
-      				<li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_Diet" />
-      				<li Class="PrisonLabor.Core.AI.JobGivers.JobGiver_Labor" />
-      			</subNodes>
-      		</li>
-      	</subNodes>
-      </thinkRoot>
-  </ThinkTreeDef>
+        </thinkRoot>
+    </ThinkTreeDef>
 </Defs>
