+++ Source/Core/AI/ThinkNodes/ThinkNode_JobGiver_PickupWeapon.cs
