+++ Source/Core/AI/WorkGivers/WorkGiver_Supervise.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using PrisonLabor.Core.Needs;
+using PrisonLabor.Core.Other;
 using PrisonLabor.Core.Trackers;
 using RimWorld;
 using Verse;
@@ -11,20 +13,27 @@ namespace PrisonLabor.Core.AI.WorkGivers
         public override Job JobOnThing(Pawn pawn, Thing t, bool forced = false)
         {
             var prisoner = t as Pawn;
-            var need = prisoner?.needs.TryGetNeed<Need_Motivation>();
-            if (need == null || prisoner == null)
-                return null;
-            if (!ShouldTakeCareOfPrisoner(pawn, prisoner))
-                return null;
-            if (prisoner.Downed || !pawn.CanReserve(t, 1, -1, null, false) || !prisoner.Awake())
-                return null;
+            var mNeed = prisoner?.needs.TryGetNeed<Need_Motivation>();
+
             if (pawn.IsPrisoner)
                 return null;
+
+            if (mNeed == null || prisoner == null || prisoner.Dead)
+                return null;
+
+            if (!ShouldTakeCareOfPrisoner(pawn, prisoner))
+                return null;
+
+            if ((!prisoner.InBed() && prisoner.Downed) || !pawn.CanReserve(t, 1, -1, null, false) || !prisoner.Awake())
+                return null;
+
+            if (prisoner.needs.food.CurLevelPercentage < 0.25 && !PrisonFoodUtility.FoodAvailableInRoomTo(prisoner) && prisoner.guest.CanBeBroughtFood && pawn.jobs.curJob == null)
+                new Job(DefDatabase<JobDef>.GetNamed("PrisonLabor_PrisonerDeliverFoodSupervise"), prisoner);
 
             if (PrisonLaborUtility.RecruitInLaborEnabled(prisoner))
                 return new Job(JobDefOf.PrisonerAttemptRecruit, t);
 
-            if ((!PrisonLaborUtility.WorkTime(prisoner) || !need.ShouldBeMotivated))
+            if ((!PrisonLaborUtility.WorkTime(prisoner) || !mNeed.ShouldBeMotivated))
                 return null;
 
             return new Job(DefDatabase<JobDef>.GetNamed("PrisonLabor_PrisonerSupervise"), prisoner);
